+++ examples/java/com/ibm/watson/developer_cloud/retrieve_and_rank/v1/ResizeASolrClusterExample.java
@@ -1,23 +1,21 @@
-/**
- * Copyright 2016 IBM Corp. All Rights Reserved.
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 package com.ibm.watson.developer_cloud.retrieve_and_rank.v1;
 
-import retrieve_and_rank.v1.RetrieveAndRank;
-import retrieve_and_rank.v1.model.SolrCluster;
-import retrieve_and_rank.v1.model.SolrCluster.Status;
-import retrieve_and_rank.v1.model.SolrClusterOptions;
-import retrieve_and_rank.v1.model.SolrClusterSizeResponse;
+import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.model.SolrCluster;
+import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.model.SolrCluster.Status;
+import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.model.SolrClusterOptions;
+import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.model.SolrClusterSizeResponse;
 
 /**
  * Example of how to create and resize a {@link SolrCluster} with {@link RetrieveAndRank}.
@@ -41,8 +39,7 @@ public class ResizeASolrClusterExample {
     }
     // 3 resize the cluster form size 1 to size 2
     final int desiredClusterSize = 2;
-    SolrClusterSizeResponse response =
-        service.resizeSolrCluster(solrClusterId, desiredClusterSize).execute();
+    SolrClusterSizeResponse response = service.resizeSolrCluster(solrClusterId, desiredClusterSize).execute();
     System.out.println(response.toString());
 
     // 4 wait until the resize processs is complete