+++ examples/java/com/ibm/watson/developer_cloud/tradeoff_analytics/v1/TradeoffAnalyticsExample.java
@@ -1,15 +1,14 @@
 /**
  * Copyright 2015 IBM Corp. All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 package com.ibm.watson.developer_cloud.tradeoff_analytics.v1;
 
@@ -17,12 +16,12 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-import tradeoff_analytics.v1.model.Dilemma;
-import tradeoff_analytics.v1.model.Option;
-import tradeoff_analytics.v1.model.Problem;
-import tradeoff_analytics.v1.model.column.Column;
-import tradeoff_analytics.v1.model.column.Column.Goal;
-import tradeoff_analytics.v1.model.column.NumericColumn;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.Dilemma;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.Option;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.Problem;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.column.Column;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.column.Column.Goal;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.column.NumericColumn;
 
 public class TradeoffAnalyticsExample {
 
@@ -40,8 +39,7 @@ public class TradeoffAnalyticsExample {
     List<Column> columns = new ArrayList<Column>();
     problem.setColumns(columns);
 
-    columns.add(new NumericColumn().range(0, 100).key(price).goal(Goal.MIN)
-        .objective(true));
+    columns.add(new NumericColumn().range(0, 100).key(price).goal(Goal.MIN).objective(true));
     columns.add(new NumericColumn().key(screen).goal(Goal.MAX).objective(true));
     columns.add(new NumericColumn().key(ram).goal(Goal.MAX));